backend/requirements.txt

fastapi
uvicorn
python-multipart
pillow
openai>=1.35.0
requests
google-cloud-storage
google-cloud-firestore
requests
python-jose
google-auth
google-generativeai>=0.7.0


